Prefactor CreateManifest endpoint ahead of upsert
