+++ src/IIIFPresentation/API/Features/Manifest/Requests/CreateManifest.cs
@@ -1,22 +1,8 @@
-﻿using System.Data;
-using API.Auth;
-using API.Converters;
-using API.Features.Manifest.Helpers;
-using API.Features.Storage.Helpers;
-using API.Helpers;
-using API.Infrastructure.AWS;
-using API.Infrastructure.IdGenerator;
+﻿using API.Converters;
 using API.Infrastructure.Requests;
-using API.Infrastructure.Validation;
-using Core;
-using IIIF.Serialisation;
 using MediatR;
 using Models.API.General;
 using Models.API.Manifest;
-using Models.Database.General;
-using Repository;
-using Collection = Models.Database.Collections.Collection;
-using DbManifest = Models.Database.Collections.Manifest;
 
 namespace API.Features.Manifest.Requests;
 
@@ -36,90 +22,17 @@ public class CreateManifest(
 }
 
 public class CreateManifestHandler(
-    PresentationContext dbContext,
-    IIdGenerator idGenerator,
-    IIIFS3Service iiifS3,
-    ILogger<CreateManifestHandler> logger) : IRequestHandler<CreateManifest,
+    ManifestService manifestService) : IRequestHandler<CreateManifest,
     ModifyEntityResult<PresentationManifest, ModifyCollectionType>>
 {
-    public async Task<ModifyEntityResult<PresentationManifest, ModifyCollectionType>> Handle(CreateManifest request,
+    public Task<ModifyEntityResult<PresentationManifest, ModifyCollectionType>> Handle(CreateManifest request,
         CancellationToken cancellationToken)
     {
-        var parentCollection = await dbContext.Collections.Retrieve(request.CustomerId,
-            request.PresentationManifest.GetParentSlug(), cancellationToken: cancellationToken);
+        var upsertRequest = new UpsertManifestRequest(request.CustomerId, 
+            request.PresentationManifest,
+            request.RawRequestBody,
+            request.UrlRoots);
 
-        var parentErrors = ValidateParent(parentCollection, request.PresentationManifest, request.UrlRoots);
-        if (parentErrors != null) return parentErrors;
-
-        var (error, dbManifest) = await UpdateDatabase(request, parentCollection!, cancellationToken);
-        if (error != null) return error; 
-
-        await SaveToS3(dbManifest!, request, cancellationToken);
-        
-        return ModifyEntityResult<PresentationManifest, ModifyCollectionType>.Success(
-            request.PresentationManifest.SetGeneratedFields(dbManifest!, request.UrlRoots), WriteResult.Created);
-    }
-
-    private static ModifyEntityResult<PresentationManifest, ModifyCollectionType>? ValidateParent(
-        Collection? parentCollection, PresentationManifest manifest, UrlRoots urlRoots)
-    {
-        if (parentCollection == null) return ErrorHelper.NullParentResponse<PresentationManifest>();
-        if (!parentCollection.IsStorageCollection) return ManifestErrorHelper.ParentMustBeStorageCollection<PresentationManifest>();
-        if (manifest.IsUriParentInvalid(parentCollection, urlRoots)) return ErrorHelper.NullParentResponse<PresentationManifest>();
-
-        return null;
-    }
-    
-    private async Task<(ModifyEntityResult<PresentationManifest, ModifyCollectionType>?, DbManifest?)> UpdateDatabase(
-        CreateManifest request, Collection parentCollection, CancellationToken cancellationToken)
-    {
-        var id = await GenerateUniqueId(request, cancellationToken);
-        if (id == null) return (ErrorHelper.CannotGenerateUniqueId<PresentationManifest>(), null);
-
-        var timeStamp = DateTime.UtcNow;
-        var dbManifest = new DbManifest
-        {
-            Id = id,
-            CustomerId = request.CustomerId,
-            Created = timeStamp,
-            Modified = timeStamp,
-            CreatedBy = Authorizer.GetUser(),
-            Hierarchy =
-            [
-                new Hierarchy
-                {
-                    Slug = request.PresentationManifest.Slug!,
-                    Canonical = true,
-                    Type = ResourceType.IIIFManifest,
-                    Parent = parentCollection.Id,
-                }
-            ]
-        };
-        dbContext.Add(dbManifest);
-        var saveErrors =
-            await dbContext.TrySave<PresentationManifest>("manifest", request.CustomerId, logger, cancellationToken);
-
-        return (saveErrors, dbManifest);
-    }
-
-    private async Task<string?> GenerateUniqueId(CreateManifest request, CancellationToken cancellationToken)
-    {
-        try
-        {
-            return await dbContext.Manifests.GenerateUniqueIdAsync(request.CustomerId, idGenerator, cancellationToken);
-        }
-        catch (ConstraintException ex)
-        {
-            logger.LogError(ex, "Unable to generate a unique manifest id for customer {CustomerId}",
-                request.CustomerId);
-            return null;
-        }
-    }
-    
-    private async Task SaveToS3(DbManifest dbManifest, CreateManifest request, CancellationToken cancellationToken)
-    {
-        var iiifManifest = request.RawRequestBody.FromJson<IIIF.Presentation.V3.Manifest>();
-        await iiifS3.SaveIIIFToS3(iiifManifest, dbManifest, dbManifest.GenerateFlatManifestId(request.UrlRoots),
-            cancellationToken);
+        return manifestService.Upsert(upsertRequest, cancellationToken);
     }
 }
